contentskit/ai-services/whisper-asr-webservice/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
numpy==1.26.4
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
--find-links https://download.pytorch.org/whl/torch_stable.html 